Update - WIP

+++ crates/db_schema/src/impls/site.rs
@@ -4,8 +4,8 @@ use url::Url;
 
 impl Crud for Site {
   type Form = SiteForm;
-  type IdType = i32;
-  fn read(conn: &PgConnection, _site_id: i32) -> Result<Self, Error> {
+  type IdType = i64;
+  fn read(conn: &PgConnection, _site_id: i64) -> Result<Self, Error> {
     use crate::schema::site::dsl::*;
     site.first::<Self>(conn)
   }
@@ -15,13 +15,13 @@ impl Crud for Site {
     insert_into(site).values(new_site).get_result::<Self>(conn)
   }
 
-  fn update(conn: &PgConnection, site_id: i32, new_site: &SiteForm) -> Result<Self, Error> {
+  fn update(conn: &PgConnection, site_id: i64, new_site: &SiteForm) -> Result<Self, Error> {
     use crate::schema::site::dsl::*;
     diesel::update(site.find(site_id))
       .set(new_site)
       .get_result::<Self>(conn)
   }
-  fn delete(conn: &PgConnection, site_id: i32) -> Result<usize, Error> {
+  fn delete(conn: &PgConnection, site_id: i64) -> Result<usize, Error> {
     use crate::schema::site::dsl::*;
     diesel::delete(site.find(site_id)).execute(conn)
   }
